pyproject.toml
[tool.poetry]
name = "searchmodel"
version = "0.1.0"
description = "Реализация простой поисковой модели на примере сервиса кинопоиск."
authors = ["mmigur <mmigur@bk.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
django-environ = "^0.11.2"
distutils-pytest = "^0.2.1"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
